requirements.txt
pytube
discord.py
python-dotenv
discord.py[voice]
PyNaCl
hugchat
